requirements.txt
streamlit
scikit-learn
xgboost
tensorflow
pandas
numpy
joblib
pytesseract
pillow
google-generativeai
deep-translator
langchain
langchain-community
langchain-google-genai
faiss-cpu
sentence-transformers
shap
fpdf2